+++ packages/components/src/components/chip/chip.stories.ts
@@ -4,6 +4,8 @@ import { boolean, modesDarkDefault, optionalAttribute } from "../../../.storyboo
 import { html } from "../../../support/formatting";
 import { ATTRIBUTES } from "../../../.storybook/resources";
 import { Chip } from "./chip";
+import "./chip"; // Force Vite to statically trace the file for Chromatic's TurboSnap feature
+import "../avatar/avatar"; // Force Vite to statically trace the file for Chromatic's TurboSnap feature
 
 const { scale, appearance, kind } = ATTRIBUTES;