+++ packages/components/src/components/link/link.stories.ts
@@ -3,6 +3,7 @@ import * as icons from "../../../../../node_modules/@esri/calcite-ui-icons";
 import { html } from "../../../support/formatting";
 import { ATTRIBUTES } from "../../../.storybook/resources";
 import { Link } from "./link";
+import "./link"; // Force Vite to statically trace the file for Chromatic's TurboSnap feature
 
 // we can get all unique icon names from all size 16 non-filled icons.
 const iconNames = Object.keys(icons)